requirements.txt
pdfplumber==0.11.9
pymongo==4.16.0
python-dotenv==1.2.2
